rename view manager file
