+++ packages/view/src/index.ts
@@ -1,6 +1,6 @@
 
 export { HandlebarsCompiler } from './engines/index.js'
-export { ViewManager } from './manager.js'
+export { ViewManager } from './view-manager.js'
 export { ViewConfigBuilder } from './view-config-builder.js'
 export { View } from './view-response.js'
 export { ViewServiceProvider } from './view-service-provider.js'
